Thu, May 18, 2023, 5:59 PM +02:00

+++ week02/hours/hours.c
@@ -42,6 +42,6 @@ float calc_hours(int hours[], int weeks, char output)
 
     else
     {
-        return sum / weeks;
+        return (float) sum / weeks;
     }
 }
